Evas textgrid: Fixed compilation warning.
SVN revision: 78360

+++ legacy/evas/src/lib/canvas/evas_object_textgrid.c
@@ -323,7 +323,7 @@ evas_object_textgrid_textprop_unref(Evas_Object_Textgrid *o, unsigned int props_
      {
         if (props->info->refcount == 1)
           eina_array_push(&o->glyphs_cleanup,
-                          (void *)((unsigned long)props_index));
+                          (void *)((uintptr_t)props_index));
         else
           evas_common_text_props_content_unref(props);
      }
